# docs/openapi/v3.yaml
openapi: 3.1.0
info:
  title: TRAE API
  description: TRAE OpenAPI 3.1 Specification
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.trae.example.com
    description: Production server
components:
  schemas:
    Problem:
      description: RFC 7807 Problem Details
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Shelf:
      type: object
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        bookCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    SearchParam:
      name: search
      in: query
      description: Search query string
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication using JWT token
security:
  - BearerAuth: []
paths:
...
  /api/v1/shelves:
    get:
      summary: List all shelves
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of shelves
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Shelf'
                  total:
                    type: integer
                    minimum: 0
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create new shelf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '201':
          description: Shelf created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelf'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/shelves/{id}:
    get:
      summary: Get shelf by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shelf details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelf'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update shelf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          description: Shelf updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelf'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete shelf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Shelf deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...